chore: separate version commands
