+++ src/command.ts
@@ -1,15 +1,10 @@
 import * as path from 'path'
 import env from './env'
-import {checkNever, throws, Version} from './utils'
+import {checkNever, throws} from './utils'
 
 function println(body: string): void {
   // eslint-disable-next-line no-console
   console.log(body)
-}
-
-function eprintln(body: string): void {
-  // eslint-disable-next-line no-console
-  console.warn(body)
 }
 
 type Command =
@@ -18,19 +13,7 @@ type Command =
   | ['set-git-user', string]
   | ['get-version']
   | ['set-version', string]
-  | ['version-unsnapshot', string] // deprecated
-  | ['version-stable', string]
-  | ['version-snapshot', string]
-  | ['version-alpha', ...([ver: string, num: string] | [ver: string])]
-  | ['version-beta', ...([ver: string, num: string] | [ver: string])]
-  | ['version-candidate', ...([ver: string, num: string] | [ver: string])]
-  | ['version-major', string]
-  | ['version-minor', string]
-  | ['version-patch', string]
-  | ['version-get-channel', string]
-  | ['version-set-channel', ...([ver: string, channel: string, num: string] | [ver: string, channel: string])]
-  | ['version-next', ...([ver: string] | [ver: string, channel: string])]
-  | ['version-format', string]
+  | import("./commands/version-commands").VersionCommand
   | ['generate-changelog', ...string[]]
   | ['prepare-gradle-maven', string, ...string[]]
   | ['prepare-gradle-signing', string, ...string[]]
@@ -110,149 +93,20 @@ async function mainImpl(...args: Command): Promise<void> {
       await changers.setVersionName(version)
       break
     }
-    case 'version-unsnapshot': {
-      eprintln("version-unsnapshot is deprecated. use version-stable")
-      println(Version.parse(args[1]
-        ?? throws(new Error('version name required')))
-        .makeStable()
-        .toString())
-      break
-    }
-    case 'version-stable': {
-      println(Version.parse(args[1]
-        ?? throws(new Error('version name required')))
-        .makeStable()
-        .toString())
-      break
-    }
-    case 'version-snapshot': {
-      println(Version.parse(args[1]
-        ?? throws(new Error('version name required')))
-        .makeSnapshot()
-        .toString())
-      break
-    }
-    case 'version-alpha': {
-      println(Version.parse(args[1]
-        ?? throws(new Error('version name required')))
-        .makeAlpha(parseInt(args[2] ?? '1'))
-        .toString())
-      break
-    }
-    case 'version-beta': {
-      println(Version.parse(args[1]
-        ?? throws(new Error('version name required')))
-        .makeBeta(parseInt(args[2] ?? '1'))
-        .toString())
-      break
-    }
-    case 'version-candidate': {
-      println(Version.parse(args[1]
-        ?? throws(new Error('version name required')))
-        .makeCandidate(parseInt(args[2] ?? '1'))
-        .toString())
-      break
-    }
-    case 'version-major': {
-      println(Version.parse(args[1]
-        ?? throws(new Error('version name required')))
-        .makeMajorOnly()
-        .toString())
-      break
-    }
-    case 'version-minor': {
-      println(Version.parse(args[1]
-        ?? throws(new Error('version name required')))
-        .makeMajorMinor()
-        .toString())
-      break
-    }
-    case 'version-patch': {
-      println(Version.parse(args[1]
-        ?? throws(new Error('version name required')))
-        .makeMajorMinorPatch()
-        .toString())
-      break
-    }
-    case 'version-get-channel': {
-      println(Version.parse(args[1]
-        ?? throws(new Error('version name required')))
-        .release[0])
-      break
-    }
-    case 'version-set-channel': {
-      let version = Version.parse(args[1]
-        ?? throws(new Error('version name required')))
-      switch (args[2].toLowerCase()) {
-        case 'a':
-        case 'alpha':
-        case 'α':
-          version = version.makeAlpha(parseInt(args[3] ?? '1'))
-          break
-        case 'b':
-        case 'beta':
-        case 'β':
-          version = version.makeBeta(parseInt(args[3] ?? '1'))
-          break
-        case 'rc':
-        case 'candidate':
-          version = version.makeCandidate(parseInt(args[3] ?? '1'))
-          break
-        case 'snapshot':
-          version = version.makeSnapshot()
-          break
-        case 'stable':
-          version = version.makeStable()
-          break
-        default:
-          throw new Error(`unknown release channel: ${args[2]}`)
-      }
-      println(version.toString())
-      break
-    }
-    case 'version-next': {
-      let target: "prerelease" | "patch" | "minor" | "major" | null
-      switch (args[2]) {
-        case null:
-        case undefined:
-          target = null
-          break
-        case "pre":
-        case "prerelease":
-        case 'a':
-        case 'alpha':
-        case 'α':
-        case 'b':
-        case 'beta':
-        case 'β':
-        case 'rc':
-        case 'candidate':
-        case 'snapshot':
-          target = "prerelease"
-          break
-        case "pat":
-        case "patch":
-          target = "patch"
-          break
-        case "min":
-        case "minor":
-          target = "minor"
-          break
-        case "maj":
-        case "major":
-          target = "major"
-          break
-        default:
-          throw new Error(`unknown next version target: ${args[2]}`)
-      }
-      println(Version.parse(args[1]
-        ?? throws(new Error('version name required')))
-        .next(target)
-        .toString())
-      break
-    }
+    case 'version-unsnapshot':
+    case 'version-stable':
+    case 'version-snapshot':
+    case 'version-alpha':
+    case 'version-beta':
+    case 'version-candidate':
+    case 'version-major':
+    case 'version-minor':
+    case 'version-patch':
+    case 'version-get-channel':
+    case 'version-set-channel':
+    case 'version-next':
     case "version-format": {
-      println(Version.parse(args[1] ?? throws(new Error('version name required'))).toString())
+      await (await import("./commands/version-commands")).runVersionCommands(args)
       break
     }
     case 'generate-changelog': {
